from Bret S. Lambert:
Included for your perusal is a monstro diff to yank the extraction of mbuf
data in NFS forward by about 15 parsecs. It's big, it's ugly, and it makes
me happy to see so much goo removed in one go.

If people would like, I can break this up, but I figure that it might as
well whack out a bunch of crap all in one fell swoop.

This also shrinks my amd64 kernel by ~7K.

- Bert