from Theo..
We have a sleep in an uncomfortable place in the midst of kern_fork
due to systrace suddenly deciding that it needs memory.

This masssages the systrace code so that it has a systrace_getproc()
which pre-allocates the two memory pieces needed.  Then later on if we
do not use systrace, we can throw that memory away using
systrace_freeproc().  While adding these functions I made systrace use
them itself too, which simplifies things a fair bit.  I also had to
use PR_ZERO instead of manual bzero'ing in the wrong places (which
early on caused me to NULL a pointer and leak memory).

I think this will now work, but I would like some tests from people
who use systrace.