bridge(4) is nice to glue qemus together and for other freaky setups. Normaly you end up with one fully configured interface while the others are just upped. This works until you hit multicast traffic. Multicast is per interface specific but bridge(4) only sent the packet up the stack with the original recieve interface (so normaly the wrong one). Because of this multicast traffic was not seen across a bridge(4).
While bridge(4) does a broadcast for multicast packets it will only send
out the packets but forgets to pass them up the stack as well.
I created a function that will do that so multicast packets are sent on
all local pathes to the network stack.

This solves my problem with qemu, brigde and ospf and I think this will
also solve the few IPv6 related troubles that where seen on the public
mailing-lists.

Maybe we should call bridge_localbroadcast() only for M_MCAST packets and
not for broadcast ones (one copy should be fine for those).

Comments, test reports, whatever?