From Henning Brauer..
THIS WILL NOT MAKE RELEASE.
DO NOT STOP DOING RELEASE CRITICAL WORK FOR THIS!

this is my monster diff, mostly from basel. it does two things,
mostly, and will likely be commited in two parts as well.

1) match rules
match rules differ from pass and block rules insofar that they
a) do not modify pass/block status of a matched packet and
b) actions* on them are always executed.
*actions: for now, queue assignment, rtable assignment, logging
means:
 pass
 match log on fxp0
passes the packet, even tho the "match" rule was the last matching
one, kinda. logging is performed.
  block
  match log on fxp0
blocks the packet and logging is done.
  match log(to pflog1) proto tcp to port 25
  pass log proto tcp to port 25
performs logging twice, once to pflog0 (the default, from the pass
rule) and pflog1 (from the match rule).

2) eleminates the scrub ruleset and scrub rules alltogether
scrub has a design failure (which is explainable from the history, but
anyway): it mixed two things that should not have been mixed, fragment
reassembly and the other packet scrubbing.
packet reassembly has to be done before state matching and ruleset
evaluation, obviously. this is the reason we traverse the scrub
ruleset 4 times (!) per forwarded packet.
the other things scrub does can be done stateful. so this is what this
diff does:
-split reassembly. after talking to frantzen, we agreed that the only
reassembly algorithm of value today is the full one, so the other two
(crop and drop-ovl) just go away. also, there is no need for a
ruleset, reassembly is either on or off. no-df is still relevant
because this controls wether fragments with the df flag set can enter
the fragment cache (after clearing df)
  set reassemble yes|no [no-df]
(in a seperate step, we really want to make reassembly on by default)

-scrub becomes a rule option, just like keep state, route to and friends.
match scrub(no-df, min-ttl 64, max-mss 1000, set-tos lowdelay)
match scrub(reassemble tcp, random-id)
(you see why we want "match" for this, and of course the list of
actions always applied above is extended by "scrub")

i have not benched yet, but this should give something in the 10%
range. plus it is just so much more clean.

why am I sending this out?
-feedback
-testing (don't let this interfere with release testing!!!)
-the manpage needs a major overhaul...
-regress tests too

and in case someone did miss it - the plan is that all forms of NAT go
the same way as scrub, i. e. become actions on rules. we'll be able to
do rdr out and nat in and nat and rdr at the same time and even
multiple nat/rdrs and everything. it'll be cool :) and match is
the logical prerequisite for all of this.