Basically, if you have the following as your ruleset, you would likely assume that only icmp-echo type traffic will pass through your firewall.
        block icmp
        pass proto icmp icmp-type echoreq

However, it turns out that an attacker who can create a state with a
ping packet can then push in any other non-session related ICMP packets
he'd like. ie. timestamp requests, router solicitations/advertisements,
network mask requests, etc.

So in the diff below, we get rid of the redundant icmp_id in the state
key, and instead store the icmp_id as one 'port', and a 'virtual' icmp
type in the other, so that request/reply pairs will only match against
that specific state.

This is very tricky code to work with, and the diff is fairly long, so
I'm very concerned about introducing new bugs right before the release.
We need people to read this carefully and really test.