Here is a much better and slightly stricter version of the diff that doesn't do any weird ICMP special casing outside of the ICMP path (that is to say it does the same weird port swapping as all the other protocols)
In order to do this, we make the cross-protocol state matching work
strictly on the pure query/response ICMP types, and use the icmp type as
a hint for which way to flip the 'ports' when we do a lookup; All other
ICMP types will match only their own ICMP type, and need to be either
enumerated specificially, or passed with a generic 'proto icmp' rule.

This is cleaner code, the logic strangeness is contained, and it
requires no userland changes.