Same description as before. It mostly works this time.
It tries to keep processes on the same cpu, but doesn't try very hard.
I want to get this version (or something better, but not more complicated)
in and improve it in tree.

Tested and works on amd64 and sparc64. This is a change from before when
amd64 didn't work.

Please make sure that it works on your favourite architectures. I'll be
poking the right people soon to get full coverage.

//art