Subject: bufcachepercent=90 freaks - you know who you are... From: Bob Beck <beck@cvs.openbsd.org> Message-Id: <200909301424.n8UEO623011617@cvs.openbsd.org> To: tech@openbsd.org CC: oga@openbsd.org, thib@openbsd.org, art@openbsd.org
And I love you for it.

If you like playing with big giant buffer caches, please try this diff
and let me know if you have any issues with it.

This diff fixes an issue where we weren't handling the inactive pages
case correctly in the page daemon. In a nutshell, inactive pages is a
count maintained by the page daemon to know how many memory pages it
can recover cheaply. they are pages that have been marked as
reclaimable by the clock hand algorithm in the pager. We weren't dealing
with this correctly wrt backing off the buffer cache, and accounting
for pages we could reclaim cheaply, which now includes the buffer cache,
the result being that we could sit there and swap with the buffer cache
holding on to memory it should release.

This could use some relatively large exposure on all arches, and on large
and small memory machines. Please let me know if you encounter any issues
with it.

Thanks.

-Bob