Doing my bit to keep dlg sane, finish off the long standing work to eliminate TRY_AGAIN_LATER in favour of NO_CCB. This will allow him to proceed with much easier to understand SCSI mid-layer rewriting.
These should be mechanical in nature. The difference in function
is that i/o's that cannot be started will be pushed back up to the
originator rather than being re-tried inside the scsi mid-layer.
For block i/o they will be requeued and retried. For others it will
be up to the originator to react.

All other SCSI drivers already do this, having adopted NO_CCB over
the last couple of years.

I want to get this in quickly to give it some days at least before
being at h2k9 and able to deal with any weirdness that results.

ok? Objections? Feel free to ok/object to particular drivers.

.. krw@