We have three knobs that you need to test with all possible combinations. We need to know which one work on your laptop and what's the bios string for the given laptop. You can look in the file for examples.
The way this goes is:
        - you apply the diff, compile and boot the kernel
        - once back at a shell you start setting each combination from
ddb
        - after you have all three variables set the way you want
suspend and resume
        - if it works add it to the okay cobinations in the file, you
may have multiple entries per laptop I've seen it happen

The variables are:
        1) acpi_repost_video
        2) acpi_clean_registers
        3) do_pci_vga_post

They are all 1 by default.

After you test just want your entries in the file, I want as little mail
in my inbox as possible. Just write if it doesn't work in any way for
you or if something panics because of these changes.

Thank you for testing.
.. pirofti@