Hi guys,
The following diff avoids the current racing on suspend (between X and
apmd) by taking X completely out of the equation. Is came about after
discussions with theo. I noted afterwards that linux actually does a
similar thing though.

Specifically, on suspend, we check to see if any of our wsdisplays are
on a vt that is in MODE_MAPPED (but not MODE_DUMBFB), and if so we
switch them to a non-MAPPED vt in the kernel. Note that if/when we get
kernel modesetting then they also shouldn't need this behaviour, since
like dumbfb there is nothing for userland to worry about.

This can't be a normal suspend hook since we may need userland to do
things for us after it gets a signal, hence the MD bits.  I have done
all archs that are capable of suspend right now, including acpi. I can
update my zaurus to test this hopefully in the next few days, but if
someone is in a position to do this sooner, please do so.  Also, if you
have a machine that actually works properly with acpi suspend, please
also see if this helps you with X. Generally, any machine that suspends
and runs X should test this diff.

With this I can not get my X40 to misbehave on suspend anymore.

The diff is a little strange since it includes a diff in src and a quick
hackish diff in xenocara to make the xserver ignore apm events (I am
thinking about a cleaner way that just ifdefing out the code). So please
beware of that.

-0-