The acpi interrupt comes in at IPL_TTY, as you can see in acpi_machdep.c.  This drives some of the back end, which then causes the top level acpi thread to advance other state.  In any case, we still have big lock to prevent one process from whacking another, and I see no point in using anything fancy here until we have all the parts needed.
So, for now, mutual exclusion of the apm top-level (open, close,
ioctl) code should be using spltty().

It's just so much better than ACPI_LOCK().  I think all those calls
were already in the right place, not that it causes much harm if
some of them are wrong.
(from Theo)