So, the problem that the mail on misc was hitting was sleeping with a mutex.
Now, turns out that this lock need not be a mutex in the first place.
We're messing with the kernel_map, non-interrupt safe. and a quick check
says that yes, we are only called from process context.

So switch the lock to a rwlock, which we may sleep with.

we also could actually do the km_free outside of the lock, having
removed the vaddr from the segment and saved it, but this diff should be
sufficient (we don't need to be at IPL_VM either way).

Added a few splasserts to make sure that we are never called at the
wrong way.

Comments, ok?

-0-