The diff below, all 3176 lines of it, fix the links between processes, pgrps, and sessions.  This will let more of the basic system calls actually work correctly with rthreads.  e.g., a child fork()ed by one thread can be wait()ed for by another; setpgrp() changes the entire process and not just the thread that called it; if a process has a controlling terminal, then *all* the threads have a controlling terminal; etc.
I've been running this in close versions since the 1st of July.  I believe
there was at least one snapshot with it; I've rerolled the diff now that
allproclk has been backed out so the diff should apply cleanly to an
up-to-date tree.  I'm looking for reports of whether anything breaks after
you apply this.  It doesn't matter whether you use rthreads, please test
this: we need confidence that committing this isn't a step backwards on
some strange case I've missed.

I won't get into the full details, but basically the change involves
moving a pile of stuff from struct proc to struct process, changing some
other stuff to point to a process instead of a proc, and then chasing
pointers in a bunch of places.  This is the first step of several.  Bret
and I tried a few times to do this particular set of changes as smaller
steps and it was impossible: the structures involved are just too
interwoven to move the links one by one.  So we get a single, 3000+ line
diff to choke down this change.

Reviews would be nice too, but mainly what this needs is to be on running
systems, proving that it behaves the same for the stuff we do all the
time.

Philip