> Sending this to a wider audience, including amd64 parts now.  This > might fix the mouse warping issues some people are still seeing.
Here is a better diff to solve the isse.  This should make the FPU
IPIs only save/flush the state of the process for which it was
intended, instead of blindly shooting down what state happened to be
on the FPU.

Please test, especially when you're seeing the mouse warping.

This could use a couple of additional pairs of eyes.