Subject: netisr is not a special little snowflake
So let's stop treating it like one, eh?

I promised this diff a few months ago and got around to it today, this
cuts the netisr handling completely over to a generic soft interrupt
handler and removes the special casing from MD code.

Now, on several architectures, this replaces the md invocation of
exactly the same. on others is replaces some special casing in
locore/vector.s that offends me greatly.

possible problems:

1) only compiled on sparc64, i386 and amd64 so far (works on all of
them, at least I still can use the network). So i may have missed bits
in the ones where my systems are either out of date (my bad) or i don't
have access to the hardware yet.

2) according to kettenis machine/intr.h is the header for
softintr_establish, I am not wholling sure of that, and if someone knows
the correct answer, please inform me and i'll fix that.

3) i'm not sure about adding a new file for this, but wasn't sure about
where it properly belonged. Again, input on this would be much
appreciated.

4) where to do the init, i put it with other network stuff, but it
doesn't really need the splnet, again, advice please?

5) copyright, since this implementation is identical to many files i'm
not sure about that either, right now it contains my copyright, but i'm
happy to deassert it if incorrect (if it ended up in an existing file
i'd not have added copyright at all).

6) the .global stuff in some .s files seems to vary between archs. I
won't pretend to be an oracle here so correct ones i've fluffed up (hppa
i'm particularly unsure about).

So, consider this a diff up for bikeshedding. It removes more lines than
it adds, kills a bunch of XXX comments and makes a bunch of things a lot
nicer.

Your comments, please

-0-