When we initialize the FPU when a process first uses it, we only initialize the FPU control registers.  This means that the FPU data registers are marked as being empty, but will still contain whatever the previous process left around in them.  It is fairly easy to dump these values to a piece of memory and look at them.  Leaking information like this is bad, but if we're going to do crypto in the kernel using the aesni instructions, it will become a major security issue.
You can actually witness this by attach gdb to a process, let it
execute a single floating point instruction, and then use the "info
float" command to look at the FPU state.  You will see that there are
many floating point registers on the register stack that are marked as
"Empty" but still contain half-sensible values.

The diff below fixes the problem.  Had it on my machine for a while
and didn't notice any problems.  However, that is a machine with SSE,
and this also needs to be tested on a machine without.  So if you have
a machinbe that doesn't have the "FXSR" feature on the cpu0: line in
the dmesg, please give this a spin for me.