Two things. DEBUG doesn't work since two return statements lack a return value. :-).
I get a failure to reset the device when DEBUG compiles. All three
conditions (res, resp[0] != PMS_RSTDONE and resp[1] != 0) seem to
be true.

When I revert r1.4 I get the mouse lockup. r1.4 moves the one mode
change call to always not polling.

So this diff keeps it working for me. Does it help yours?

.... Ken