fix various TCP issues
The new tcp send/recv socket buffer scaling uncovered a few bugs in our
stack that seem to be around for some time but were well hidden.

First hunk removes a totaly incorrect call of tcp_rscale(). It is not
allowed to recalculate the receive scaling value after the initial SYN
packet. We need to stick to the value the syncache calculated.
Because of this our window updates are totaly wrong which results in a lot
of "data after window" packets.

The second part has todo with the syncache not properly initializing the
initial TCP timestamp and by that causing havok in 50% of all connections
(because of the timestamp modulation). This results in non working receive
side buffer scaling and probably other nasty effects.
Just to be sure I also make sure that we do not use stack garbage in the
dummy tcpcb we pass to tcp_dooptions() in syn_cache_add().

This problems can be triggered by using tcpbench over the loopback.
--
:wq Claudio