more assertwaitok() love
Try to catch more places where we sleep and are not allowed.

One thing of note, msleep() is missing in this diff, but there
it is needed to call to sleep_setup routines with the mutex
held, and after we release it we _will_ sleep so a sleep there
with another mutex held will be caught by the assertwaitok()
in mi_switch().

Also, define assertwaitok() out for !DIAGNOSTIC kernels.
.. thib@