when a usb device is removed from a port, a hub interrupt is generated.  this causes a hub explore task to be scheduled, which will find that the device has been removed then detach the driver via usb_disconnect_port().  so there is some time between when the device is removed and the driver detachment. if the kernel tries to access the device after it has been removed but before the driver is detached, a kernel crash is likely. also, the kernel does not know the device has been removed until the driver is detached; config_deactivate() is never called.
the following patch adds a 'dying' field to struct usbd_device,
and a couple functions, usbd_deactivate() and usbd_is_dying().
usbd_deactivate() sets the new dying field, and usbd_is_dying()
checks whether the dying field is set.  usb_disconnect_port()
is changed to call config_deactivate() for devices to be detached
before calling config_detach(), and usbd_deactivate() is added
to the DVACT_DEACTIVATE case for many devices' activate()
functions.  also, usbd_is_dying() is either added or replaces
driver specific state variables at places where it would be
good to know if the hardware is still usable.

the idea is to give drivers a chance to quit accessing the
hardware before the driver is detached, so we don't detach
the driver while a request is stalled trying to read/write
nonexistent hardware.

oh, usbd_is_dying() also checks if the underlying bus is
detached, because if the bus is detached, the device has
also been detached.  I think this will avoid the crash
described in
http://marc.info/?l=openbsd-bugs&m=129050458925460&w=2
it seems that that device generates a hub interrupt (the
only thing that causes uhub_explore() to be run) when
detached.  on the one hand this makes sense: hub interrupt
means the status of the hub changed.  on the other, it's
nonsensical: the hub is gone, we don't need an interrupt to
know that the hub state has changed.

the idea is really from yuo@, in some commit messages from
a couple months ago.

I've been running this diff for quite some time, and it was
actually in snapshots for some time in early October.

thoughts?

p.s. yes, this doesn't convert *all* drivers at this time.
that will happen, I promise.

.. jakemsr@