Hi,
here's an updated and mostly rewritten version of the diff started a
while ago by the Ecdysis group (Simon Perreault, et al.) to provide
a free proof of concept implementation of the draft NAT64 specification.
Unfortunately their work was not finished, is not up-to-date has
limited functionality and lots of bugs.  We (jsg@, reyk@ and me) spent
some time and produced a quality diff that addresses most of the
aforementioned problems.

As it was discussed in the paper[1] released by the Ecdysis we can't
translate address families the same way we translate addresses in
nat-to and rdr-to cases because address family translation is not
compatible with routing and requires packet reinjection.

We decided to address this problem by applying NAT64 rules on the
incoming interface then making a routing decision and injecting the
packet to the outgoing interface.  This has an advantage of not
requiring the packet to traverse the ip/ip6 input queue.

To make reply packets match our rules without having a WIRE side state
key on the incoming interface we preserved the original idea of having
half reversed state so that NAT64 rules have the following state key
setup:

 consider we're connecting from 10.0.1.1 to 10.0.2.1 and we decided to
 use translation rules like that (addresses are completely arbitrary):

   10.0.1.1 -> 2001::1:1
   10.0.2.1 -> 2001::2:1

 then the WIRE side state key is:

   src: 10.0.1.1   dst: 10.0.2.1

 and the STACK side state key is:

   src: 2001::2:1  dst: 2001::1:1

This makes state lookup match reply packets and have address family
translation information at hand.

The pfctl(8) grammar we conceived is based on the style of nat-to, so
if vr0 is your incoming interface then the ipv4-to-ipv6 translation
rule looks like:

  pass in on vr0 inet af-to inet6 (2001::1 2001::/96)

and in a short form, considering a 1:1 mapping of the ipv4 address
inside the last 32 octets of an ipv6 address (allowed by the /96
prefix):

  pass in on vr0 inet af-to inet6 2001::1

where 2001::1 is a *source* address of the target ipv6 capable interface.
It can also be specified in a (int0) notation.  Source and destination
addresses don't need to be on the same subnet.  1:1 (or "af-to binat")
configurations are supported to.

The same applies to the ipv6-to-ipv4 translation.  Unfortunately man page
bits for pfctl are still in development :(

pflog/pfsync changes are a compromise at the moment.

Wrt pflog: henning@ told me that he wants to see a translated packet in
the logs, not the original one, but this will require some crafty changes
in the pflog-bpf code:  consider we're about to translate a 32 byte UDP
packet (20+8+4), resulting ipv6 packet is 20 bytes bigger (40+8+4) so we
can't even fit an ipv6 header into the buffer that bpf gave us.

Wrt pfsync: dlg@ told me that he would like to have separate message ids.
This is possible, yet Theo and I decided to send an initial diff with two
if's that allow us to be fully compatible (import and export) with older
implementations (and I've just tested this setup and it works w/o any
problems).

Also this two diffs are not the only ones yet most important.  tcpdump
and systat require small changes and ftp-proxy got a serious punch in the
balls to deal with all that, but this is going to happen next week.

Of course there's much more stuff that's worth mentioning, but I don't
want to disturb anyone from watching porn for too long, so test it,
report your results, prod for changes, have fun and stay tuned.

Cheers,
Mike

Diffs are also available at:

http://team.vantronix.net/~mike/nat64.diff
http://team.vantronix.net/~mike/pfctl.diff

[1] http://www.viagenie.ca/publications/2010-03-13-asiabsdcon-nat64.pdf