The following diff is a word in progress that allows i386 boot to load a kernel from a softraid volume. There are several major cleanups that need to be made before this goes anywhere near the tree, plus it currently only knows about RAID 1. However, I've hit a couple of challenges that I could use some assistance with.
1. With the diff below i386 boot has a size of 0xc6db (i.e. just under
50KB). If I add a little bit more code and cross the 0xc700 boundary boot
hangs the machine hard. I have not yet been able to determine if this is a
hardware limitation, a bug in biosboot, a bug in boot or a bug in
installboot... if anyone has a cluebat it would be appreciated.

2. We currently have a heap size of 768KB configured for boot and it would
appear that the ELF loader will happily eat most of this. Given that I need
some memory for the softraid metadata and data structures, I've had to
increase the heap (currently to 1MB) - are there any problems with doing
this and/or will there be nasty side effects?

If you want to play with it create yourself a RAID 1 softraid volume:

fdisk -iy sd0a
fdisk -iy sd1a
bioctl -c 1 -l /dev/sd0a,/dev/sd1a softraid0

Install the boot loader:

/usr/mdec/installboot -v /usr/mdec/boot /usr/mdec/biosboot sd2

And put a kernel on it (do not fdisk the softraid volume - this case is not
yet handled correctly):

disklabel -E sd2
newfs /dev/sd2a
mount /dev/sd2a /mnt
cp /bsd /mnt
umount /mnt

You should now be able to boot from either sd0a or sd1a, have boot include
sr0 in the list of drives and be able to run 'ls' or 'boot' using this device.
Once this is working well it should not be a lot more work to get boot on
crypto working...