=== Algorithm ===
vmmap is designed to perform address space randomized allocations,
without letting fragmentation of the address space go through the roof.

The way it achieves this is by keeping a list of free memory regions.
Whenever an allocation is performed, the best-fit free space is searched
and a random space is chosen from that space and 8 (knob) spaces after.

Within that space, a random offset is chosen, such that:
- iff the free space tree has more than 128 (knob) entries,
  the offset will touch neighbouring used memory
- otherwise, the free space will be anywere (and probably cause more
  fragmentation
This way, the allocator makes the fragmentation converge to 128 entries.

=== User space spillover ===

During writing the allocator, tedu@ made a diff that allows allocating
from the brk(2) area. The idea is that this only happens if the normal
heap cannot accomodate allocations.
Therefor, the allocator maintains a second free space list, which contains
entries in brk(2) and stack areas. When allocating from these areas, the
algorithm will start "at the far end" of those areas (high to low in brk(2),
opposite end of the stack growth direction for stack).
Once allocation from the heap fails, the allocator switches over to these
secondary areas.

=== Kernel address space randomization ===

Some archs have pre-allocated pmap pages and can't handle a naive
implementation of random address selection. Because of this, for kernel
memory, the free space is clamped down and only grown when allocations fail.
This way, address space randomization is introduced to the kernel memory.

=== VMSpace tweasks ===

Because kernel memory and user memory are treated different, I need to
know which type of map I have. Maps now carry a flag if they are vmspaces,
in which case the 'User space spillover' case is selected, otherwise the
'Kernel address space randomization' kernel growth is used during allocation
failures.

Because of the flag, the allocator will automatically update the free-space
tree, if the vmspace parameters changed since the last allocation.
Also, it will update the vmspace useage values itself, rather than requiring
caller to do it. uvm_map_p() has been changed back to uvm_map() because of
this as well: the process argument was only used to get to the vmspace
anyway.

=== Guard pages ===

The current source-tree has a hackish implementation of guard pages (my bad).
This allocator allows placing guard pages in the map and will not handle
them as free pages. For a map with guard pages, if a space between
2 allocations is <= 2 pages, it will not show up in the free space.
The disadvantage of using guard pages is that it will use more vm_map_entry:
it's not possible to merge 2 entries with a guardpage between them.

=== Hint address ===

The hint address is no longer used by vm_map. Calls to vm_map_hint() are
no longer required. Userland that wants the hint to be tried first will
have to try a fixed allocation first. uvm_mmap() behaviour has been updated
to explicitly try this if the MAP_TRYFIXED flag is set.

=== TLB pressure ===

The allocator reduces fragmentation, reducing TLB pressure and requiring
less pmap pages.

=== Small fixes ===

VM_MAP_HOLE allocations now require the VM_MAP_FIXED flag set. Holes are
for space that the mmu can't manage and thus should not be placed at
a random offset.
vm_dused is updated automatically by uvm_map. Verification is available
for this case.

The old allocator would fail allocation if the address randomization
was chosen in such a way that all memory at the point to the end of memory
was in use. This was because the old allocator was a compacting allocator
that only performed a forward search. This allocator is based on a best-fit
algorithm instead.

=== Knobs ===

The allocator has a few knobs. These exist because the need for address
space randomization and the need for reduced fragmentation contradict
eachother. The knobs are chosen to select a good middle ground.
Knobs are just internal constants.