The diff below makes use of the macobio_* methods to manage the i2s GPIOs. In order to achieve this goal, instead of reading the gpio address from the openfirmware, it now retrieves the offsets from the base bus address. The result is that we don't need to re-map the GPIOs.
It also removes the customs gpio_* methods and move all the remaining
global variables in the i2s softc.

The next work will be to have a working line in/out. Please test this
diff an report any regression. So far I haven't seen any problem on my
machines.

Martin