So the short answer about buf_needva
1) bufkvm is only computed to be 10% of kva.. that hasn't changed
   (even if we have bufcachepercent=42 )

on machines with a small enough buffer cache, so that hidirtypages was
< bufkvm (by accident) we don't hit this, because the buffer cleaner
will force out dirty pages and release them, thus freeing kva

the problem happens when bufkvm is less than hidirtypages - now we
have the problem. We are then allowed to have more dirty pages than we
have kvm - a situation the cleaner would prevents. however this was
only happening by *accident* based upon what hidirtypages happened to
be (given your machine's memory size and chosen size of
bufcahcepercent. The bigger the memory, and larger the buffer cache,
the more likely this was to be the case.

What'e below is *not the right fix* - but I believe is a cheap hack to
get us by. I've talked to theo about fixing this better and I will
post-release - however, I think this is a fix that could potentially
make release and allow us to leave the buffer cache at it's current
level without risking kvm starvation.  So the fix I am looking at
testing at the moment is one of simply making hidirty and lowdirty
(which control the cleaner) be computed based on the number of kvm
mappable pages rather than the number of total pages in the buffer
cache.

If I'm right, this will prevent buf_needva deadlocks in the most
minimal way at the moment.