Date: Thu, 22 Sep 2011 23:54:20 +0200 From: Claudio Jeker <claudio@openbsd.org> Subject: nat64 diff Message-ID: <20110922215420.GA15060@diehard.n-r-g.com>
OK, I played a bit with the nat64 diff and merged it to -current.
This seems to work for me (as in I play a bit with nc(1) and ping/ping6
and TCP/UDP and ICMP seem to work).

So people interested in this (plus Henning, Alexander and Ryan -- sorry
you can not chicken out) please have a look. I would really like to get
this in soonish since it is a pain in the ass to fix the conflicts
everytime a pf.c change happens.