The diff below corrects the handling of execve and makes coredumping safe when rthreads are in use.  These are critical for being able to switch to rthreads.
The issue for execve is that it's suppose to kill all the other threads in
the process, but only if it succeeds.  The issue for coredumping is that
the coredump code assumes the thread list doesn't change behind its back,
but there's nothing to guarantee that.

So, this is done with three new functions and process/proc bits and
members to support the ability for one threads to tell all the other
threads to either just stop where they are, or unwind back to userret and
stop, or unwind back to userret and die.
  single_thread_check(p, deep)
        If another thread has requested that we stop or exit, then either
        do so (if deep==0 indicating this is a call from userret()) or
        return an error to indicate that we should return EINTR or ERESTART
  single_thread_set(p, bits, deep)
        If another thread has requested that we stop or exit, then do
        so as with single_thread_check().  Else actually tell others to
        stop or die based on 'bits': 0 == "just stop", PS_SINGLEUNWIND ==
        "unwind to userret and stop", PS_SINGLEEXIT == "unwind to userret
        and die"
  single_thread_clear(p)
        Undo a single_thread_set() that stopped threads, letting them resume.

Test reports, with or without using rthreads, and reviews or oks would be
much appreciated!