Hi,
The huge diff included below is to replace vmmap for something that
works a lot better. This is the second version I plan to commit.
I have been running this on my laptop for months and have had no
problems.

Highlights:
- the mapping and allocation (address selection) logic is split
- the allocator is in compatibility mode, meaning it ought to perform
  the same strategies (and hence contain the same bugs, yay) as the
  version that you are running in your kernel at the moment
  (this means java and various browsers won't die horribly)
- despite using the same algorithm, it is a bit^W alot faster
- it's only a 3^W 10494 line diff :P

I'll need testing on any and all architectures.

The current incarnation is:
  vmmap_sys.diff.63