This diff actually does 2 things:
1) It fixes a sitation we found and hit where disksort was trivially
able to get into a situaion where some io's would starve. This diff
includes changes to disk sort to limit us to 128 insertions *before*
an inversion, after which we will process past the inversion and
continue.  the number 128 was chosen by me and Ken based on the IO
rate of slow disks on an old sparc. and "Seemed reasonable" - but we
will need to test this in lots of places, especially "old" machines
with old disks that rely on disksort to prevent them from waving the
heads all over hell like an epileptic baboon.

 In a nutshell, the intent of the change is to try to allow large
chains of io to be built on each "ladder" while ensuring that will
will peiodically service each ladder, rather than exclusively servicing
one for a very long period.

2) It implements a high and a low water mark for writes, per bufq, so
that when a buffer has more than the high water mark of io's in
flight, writers must wait till the low water mark is reached. It then
attempts to set the hi and low water marks on queue initialization
based upon the openings of the device. Currently we try for 4 times
openings for high, and 2 times for low - the intention here is so that
devices which can process multiple operations can still get full
performance, however we don't build up an enormous debt of io's in
flight that make interactive performance awful while that is going on.

With this, we can hammer on both disks, and usb sticks, with a big
sequential IO, that does not seem to be significantly slowed, and
meanwhile I can vi a file on the same file system. So as I try it
now on the stuff I'm beating up, it's a vast improvement over current.

I'm gonna need some help with this, some eyes, and particularly some
tests, both on "problem cases" where you have seen issues related to
this before, as well as on normal machines with sd and ata disks. I'm
especially interested in knowing if it hurts "old" disks that depend
on disksort to get decent performance.

Note when you are measuring, if you are comparing to current, you
can't do a big "dd" or "cp" and just wall time it - because you will
be comparing how long it takes to fill the buffer cache - you have ot
either contrive a test for sequential io with an fsync at the end, or
watch it for when the io *actually* stops. (I got bit thinking it hurt
a lot by dd that would finish in 13 seconds, when it *really* took
20 for the IO to finish.) so watch out for "comparing" it to current.

So, can anybody help out?